src/updated_poc/requirements.txt
diffusers @ git+https://github.com/ShivamShrirao/diffusers.git@25045fde9c4281ee317703714cc0331bb97d5a65
absl-py==1.4.0
accelerate==0.16.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bitsandbytes==0.35.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
fastapi==0.92.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
google-auth==2.16.1
google-auth-oauthlib==0.4.6
gradio==3.19.1
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.2.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
orjson==3.8.6
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pydantic==1.10.5
pydub==0.25.1
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
safetensors==0.2.8
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
xformers==0.0.16
yarl==1.8.2
zipp==3.15.0